fix(reddog): linearize diagnostic secret filtering
Worker-Lane: RedDog diagnostic architect analysis

Slice: REDDOG_DAEMON_DIAGNOSTIC_ARCHITECT_ANALYSIS_PHASE1
